Fix r7 register usage in ARM syscall
In armv7 register r7 is used to hold the syscall number.
However, r7 is also an LLVM reserved register ([1]) and cannot be used
as an input register.
As a temporary workaround we use r8 as temporary input register for
holding the syscall number and then copy r8 into r7 before the actual
system call.

[1] https://github.com/rust-lang/rust/issues/85056